'one sentence per line' format
Our reviewing pipeline is relatively new, and we're continuously
exploring ways to enhance productivity for all involved. Because
of how the original transcript is formatted - a long continuous
block of text - makes it challenging to compare the before and
after versions and assess your work. To address this, we've
transitioned to a format where each sentence is on a separate line
you will observe this change in your future reviews.

While reviewing your submission, I've implemented this new
'one sentence per line' format and this commit reflects this adjustment